chore: move functions in response to PR comments
